Moved calculate_rpm() to control.c from in4073.c
